Add files via upload
In this Repository, we have python program test and practise which was carried out during the Data Science Nigeria Boot Camp for July/August 2019 Cohort. Towards the end of the Camp, we had projects and the file titled "PY GROUP PROJECT ORIGINAL" is An inventory management system for a shop owner to store and add new products to a database. Also to collate customer information.